move definitions of enc/decoder interfaces and helpers to a new module
